adding user#create
