+++ config/routes.rb
@@ -2,6 +2,7 @@ Rails.application.routes.draw do
   root 'hacker_news_stories#index'
 
   post '/login', to: 'authentication#authenticate'
+  post '/signup', to: 'user#create'
   get '/stories', to: 'hacker_news_stories#index', as: 'hacker_news_stories'
   post '/stories', to: 'hacker_news_stories#create'
   get '/stories/new', to: 'hacker_news_stories#new', as: 'new_hacker_news_story'
