+++ test/controllers/user_controller_test.rb
@@ -1,0 +1,7 @@
+require 'test_helper'
+
+class UserControllerTest < ActionDispatch::IntegrationTest
+  # test "the truth" do
+  #   assert true
+  # end
+end
